+++ main/routes/logout.py
@@ -1,8 +1,6 @@
 from .baseRoutes import jsonify, status, app, cur, conn
-from .login import jwt_required, get_jwt_identity, login
+from .login import jwt_required, get_jwt_identity, get_raw_jwt, login
 
 @app.route('/api/v1/logout', methods = ['GET'])
 def logout():
-    pass
-
-    return jsonify({"message":"You have successfully logged out."})
+    return jsonify({"message":"You have successfully logged out."}), 200
